Fix lint priority
Lint rules in the `[lints]` table are not applied in file order. This
could cause certain rules to be ignored if they were applied before the
rule that effects the group they were in. For example, if
`clippy::nursery` and `clippy::option_if_let_else` have the same
priority, then `nursery` gets applied after `option_if_let_else`,
causing the more specific rule to be ignored.

This explicitly sets `nursery` and `pedantic` to priority -1 so that
they are overridden by more specific rules (which are priority 0 by
default).